+++ src/main.js
@@ -1,4 +1,4 @@
-import { searchFilter, filterOne, filterTwo, filterTopTen } from './data.js';
+import { searchFilter, filterOne, filterTopTen } from './data.js';
 // import data from './data/lol/lol.js';
 import data from './data/ghibli/ghibli.js';
 
@@ -34,16 +34,94 @@ search.addEventListener("keydown", (key) => {
 const order = document.getElementById("order");
 order.addEventListener("change", (event) => {
   const option = event.target.value;
-  filterOne(option, container, films, img);
+  
+// FILTRO TODOS
+   if (option === "Todas") {
+    container.innerHTML = " ";
+    for (let i = 0; i < films.length; i++) {
+      container.innerHTML += img(films[i]);
+    }
+  }
+//FILTRO AZ
+
+  if (option === "Alfabetico") {
+    container.innerHTML = " ";
+    for (let i = 0; i < films.length; i++) {
+      container.innerHTML += img(films[i]);
+      //console.log(films[i].title);
+    }
+  }
+  //FILTRO DE AÑO
+  if (option === "Año") {
+    container.innerHTML = " ";
+    for (let i = 0; i < films.length; i++) { 
+      container.innerHTML += img(films[i]);
+      //console.log(films[i].release_date);
+    }
+  }
+  filterOne (films, img, option);
 });
 
 //FILTRO POR DIRECTOR
 const director = document.getElementById("director");
 director.addEventListener("change", (event) => {
   const directorOption = event.target.value;
-  filterTwo(directorOption, container, films, img);
-});
-
+//FILTRO DIRECTOR
+  if (directorOption === "todas") {
+    container.innerHTML = " ";
+    for (let i = 0; i < films.length; i++) {
+      container.innerHTML += img(films[i]);
+    }}
+    if (directorOption === "Hayao") {
+      container.innerHTML = " ";
+      const hayaoFilter = films.filter(x => (x.director === "Hayao Miyazaki"));
+      for (let i = 0; i < hayaoFilter.length; i++) {
+        container.innerHTML += img(hayaoFilter[i]);
+        //console.log(hayaoFilter[i].director);  
+      }
+    }
+    if (directorOption === "Isao") {
+      container.innerHTML = " ";
+      const isaoFilter = films.filter(x => (x.director === "Isao Takahata"));
+      for (let i = 0; i < isaoFilter.length; i++) {
+        container.innerHTML += img(isaoFilter[i]);
+        //console.log(isaoFilter[i].director); 
+      }
+    }
+    if (directorOption === "Gorō") {
+      container.innerHTML = " ";
+      const goroFilter = films.filter(x => (x.director === "Gorō Miyazaki"));
+      for (let i = 0; i < goroFilter.length; i++) {
+        container.innerHTML += img(goroFilter[i]);
+        //console.log(goroFilter[i].director);
+        
+      }
+    }
+    if (directorOption === "Hiroyuki") {
+      container.innerHTML = " ";
+      const hiroyukiFilter = films.filter(x => (x.director === "Hiroyuki Morita"));
+      for (let i = 0; i < hiroyukiFilter.length; i++) {
+        container.innerHTML += img(hiroyukiFilter[i]);
+        //console.log(hiroyukiFilter[i].director);  
+      }
+    }
+    if (directorOption === "Hiromasa") {
+      container.innerHTML = " ";
+      const hiromasaFilter = films.filter(x => (x.director === "Hiromasa Yonebayashi"));
+      for (let i = 0; i < hiromasaFilter.length; i++) {
+        container.innerHTML += img(hiromasaFilter[i]);
+        //console.log(hiromasaFilter[i].director);  
+      }
+    }
+    if (directorOption === "Yoshifumi") {
+      container.innerHTML = " ";
+      const yoshifumiFilter = films.filter(x => (x.director === "Yoshifumi Kondō"));
+      for (let i = 0; i < yoshifumiFilter.length; i++) {
+        container.innerHTML += img(yoshifumiFilter[i]);
+        //console.log(yoshifumiFilter[i].director); 
+      }
+    }
+  });
 //PAGINAS DE DATOS
 
 const imgData = (films) => {
@@ -105,7 +183,12 @@ const printFilmData = (id) => {
 
 const pageTop= document.getElementById("clickTop");
 pageTop.addEventListener("click", () => {
-  filterTopTen(container, films, img)
+  container.innerHTML = " "; 
+  for (let i = 0; i < 10; i++) {
+    container.innerHTML += img(films[i]);
+    //console.log(films[i].rt_score);
+  }
+  filterTopTen( films)
 });
 
 //PAGINA HISTORIA
